Move versions and gen related code to generations
