+++ src/ts/generations.ts
@@ -13,6 +13,18 @@ export enum Generations {
     GEN_8 = 8,
 }
 
+export interface Versions<T> {
+    [Generations.GEN_LATEST]: T;
+    [Generations.GEN_1]?: T;
+    [Generations.GEN_2]?: T;
+    [Generations.GEN_3]?: T;
+    [Generations.GEN_4]?: T;
+    [Generations.GEN_5]?: T;
+    [Generations.GEN_6]?: T;
+    [Generations.GEN_7]?: T;
+    [Generations.GEN_8]?: T;
+}
+
 const versionGroupCache: { [key: string]: Generations; } = {};
 
 export function convertGenStringToEnum(gen: string): Generations {
@@ -38,4 +50,43 @@ export async function convertVersionGroupStringToEnum(versionGroup: string): Pro
     versionGroupCache[versionGroup] = gen;
 
     return gen;
+}
+
+export function getClosestGen(versions: Versions<unknown>, gen: Generations): Generations {
+    // get the specific gen if it exists
+    if (versions[gen]) {
+        return gen;
+    }
+
+    // if not, see if we have the data for any past generations
+    const pastGenerations = Object.keys(versions).filter(
+        (version) => version !== Generations.GEN_LATEST,
+    );
+
+    if (!pastGenerations.length) {
+        return Generations.GEN_LATEST;
+    }
+
+    // get the closest higher generation
+    const closestHigherGen = pastGenerations.reduce((closest, current) => {
+        if (current > gen) {
+            return current;
+        }
+        return closest;
+    });
+
+    // if its not higher, just return the latest
+    if (!closestHigherGen || closestHigherGen < gen) {
+        return Generations.GEN_LATEST;
+    }
+
+    // get the closest lower generation
+    const closestLowerGen = pastGenerations.reduce((closest, current) => {
+        if (current < gen) {
+            return current;
+        }
+        return closest;
+    });
+
+    return closestLowerGen as Generations;
 }
